Move block selection to its own useBlockSelection hook
